requirements.txt
claude-api>=1.0.17
click>=8.0.0
rich>=12.0.0
pyyaml>=6.0